requirements.txt
a2wsgi==1.10.10
aiofile==3.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiohttp_socks==0.11.0
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
azure-ai-documentintelligence==1.0.2
azure-core==1.38.3
azure-identity==1.25.3
beautifulsoup4==4.14.3
blinker==1.9.0
brotli==1.2.0
browser-cookie3==0.20.1
caio==0.9.25
cairocffi==1.7.1
CairoSVG==2.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cloudscraper==1.2.71
cobble==0.1.4
cryptography==46.0.5
cssselect2==0.9.0
curl_cffi==0.14.0
ddgs==9.11.3
defusedxml==0.7.1
Deprecated==1.3.1
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.135.1
Flask==3.1.3
flatbuffers==25.12.19
frozenlist==1.8.0
g4f==7.3.2
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isodate==0.7.2
itsdangerous==2.2.0
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.13.0
lxml==6.0.2
lz4==4.4.5
magika==0.6.3
mammoth==1.11.0
markdownify==1.2.2
markitdown==0.1.5
MarkupSafe==3.0.3
mpmath==1.3.0
msal==1.35.1
msal-extensions==1.3.1
mss==10.1.0
multidict==6.7.1
nest-asyncio2==1.7.2
ngrok==1.7.0
nodriver==0.48.1
numpy==2.4.3
olefile==0.47
onnxruntime==1.24.3
openai==2.28.0
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
platformdirs==4.9.4
primp==1.1.3
propcache==0.4.1
protobuf==7.34.0
pycparser==3.0
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
PyJWT==2.12.1
pyparsing==3.3.2
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
python-pptx==1.0.2
python-socks==2.8.1
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SpeechRecognition==3.15.1
starlette==0.52.1
sympy==1.14.0
tinycss2==1.5.1
tqdm==4.67.3
twilio==9.10.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
wasmtime==42.0.0
webencodings==0.5.1
websockets==16.0
Werkzeug==3.1.6
wrapt==2.1.2
xlrd==2.0.2
xlsxwriter==3.2.9
yarl==1.23.0
youtube-transcript-api==1.0.3
